# refactor(parser): build with '-optimize-grammar' flag