--- a/.github/workflows/ci-build.yml
+++ b/.github/workflows/ci-build.yml
@@ -74,5 +74,5 @@ jobs:
       uses: golangci/golangci-lint-action@v2.3.0
       with:
         version: v1.29
-        args: -E gofmt,golint,megacheck,misspell
+        args: --timeout=2m0s -c .golangci.yml