--- a/pkg/parser/generate.go
+++ b/pkg/parser/generate.go
@@ -1,3 +1,3 @@
 package parser
 
-//go:generate pigeon -optimize-parser -optimize-grammar -alternate-entrypoints RawSource,RawDocument,DocumentBlock,FileLocation,LabeledListItemTerm,SpecialCharacterSubs,MarkdownQuoteAttribution,QuotedTextSubs,NoneSubs,AttributeSubs,ReplacementSubs,PostReplacementSubs,InlinePassthroughSubs,CalloutSubs,InlineMacroSubs,MarkdownQuoteMacroSubs,BlockAttributes,LineRanges,TagRanges -o parser.go parser.peg
+//go:generate pigeon -optimize-parser -optimize-grammar -alternate-entrypoints RawSource,RawDocument,DocumentBlock,FileLocation,IncludedFileLine,LabeledListItemTerm,SpecialCharacterSubs,MarkdownQuoteAttribution,QuotedTextSubs,NoneSubs,AttributeSubs,ReplacementSubs,PostReplacementSubs,InlinePassthroughSubs,CalloutSubs,InlineMacroSubs,MarkdownQuoteMacroSubs,BlockAttributes,LineRanges,TagRanges -o parser.go parser.peg