Signed-off-by: Xavier Coulon <xcoulon@redhat.com>